Split out tests, simplify models
